+++ src/index.js
@@ -7,6 +7,8 @@ import FreshnessIndicator from './components/freshness-indicator/freshness-indic
 import onClickOutside from './hocs/on-click-outside/on-click-outside';
 import IconRow from './components/icon-row/icon-row';
 
+import utils from './utils';
+
 export {
   // Number components
   Currency,
@@ -21,4 +23,7 @@ export {
 
   // HOCs
   onClickOutside,
+
+  // utils
+  utils,
 };