+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Upcoming release
 * Development - Fixed broken decimalProps.
-* Utils - Exported utility functions for getTickDecimals.
+* Utils - Exported utility functions for getTickDecimals, numberIsFinite and getFractionDigits.
 
 ## 0.2.1
 * IconRow - Fixed broken custom styling.