+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## Upcoming release
 * Development - Fixed broken decimalProps.
-* Utils - Exported utility functions for getTickDecimals, numberIsFinite and getFractionDigits.
+* Utils - Exported utility functions for getTickDecimals, numberIsFinite and getFractionDigits. Usage:
+
+  ```js
+  import { utils } from 'nordnet-component-kit';
+
+  //...
+  const decimals = getTickDecimals(value, ticks);
+```
 
 ## 0.2.1
 * IconRow - Fixed broken custom styling.