+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming release
+## 0.2.2
 * Development - Fixed broken decimalProps.
 * Utils - Exported utility functions for getTickDecimals, numberIsFinite and getFractionDigits. Usage: